We only support 64-bit systems for now
* Don't try to build 32-bit Python wheels.

* Ensure hash types are large enough.

+++ src/menai/vm/menai_vm_c.h
@@ -226,8 +226,8 @@ menai_xrelease(MenaiValue *val)
     }
 }
 
-typedef ssize_t hash_t;
-typedef size_t uhash_t;
+typedef int64_t hash_t;
+typedef uint64_t uhash_t;
 
 /*
  * menai_hash_double — hash a C double without any Python API calls.
